finally merge a patch from Kuri which adds custom emoticon support and update the ui to match this api break. also fix a bug where jid was not properly sanitized upon creating initial connection object
git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@76740 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ PROTO/shotgun/src/lib/shotgun_utils.c
@@ -3,7 +3,7 @@
 #include "cdecode.h"
 
 char *
-shotgun_base64_encode(const unsigned char *string, double len, size_t *size)
+shotgun_base64_encode(const unsigned char *string, size_t len, size_t *size)
 {
    base64_encodestate s;
    char *ret = NULL;
